move to offline package, and make compatible with new system
